#640 [refactor] Change folderId type to Long

+++ data/src/main/java/daily/dayo/data/datasource/remote/folder/FolderApiService.kt
@@ -18,7 +18,7 @@ interface FolderApiService {
     @Multipart
     @POST("/api/v1/folders/patch")
     suspend fun requestEditFolder(
-        @Part("folderId") folderId: Int,
+        @Part("folderId") folderId: Long,
         @Part("name") name: String,
         @Part("privacy") privacy: String,
         @Part("subheading") subheading: String?,
@@ -30,7 +30,7 @@ interface FolderApiService {
     suspend fun requestCreateFolderInPost(@Body body: CreateFolderInPostRequest): NetworkResponse<CreateFolderInPostResponse>
 
     @POST("/api/v1/folders/delete/{folderId}")
-    suspend fun requestDeleteFolder(@Path("folderId") folderId: Int): NetworkResponse<Void>
+    suspend fun requestDeleteFolder(@Path("folderId") folderId: Long): NetworkResponse<Void>
 
     @POST("/api/v1/folders/order")
     suspend fun requestOrderFolder(@Body body: List<EditOrderDto>): NetworkResponse<Void>
@@ -44,9 +44,9 @@ interface FolderApiService {
 
     // 폴더 정보
     @GET("/api/v1/folders/{folderId}/info")
-    suspend fun requestFolderInfo(@Path("folderId") folderId: Int): NetworkResponse<FolderInfoResponse>
+    suspend fun requestFolderInfo(@Path("folderId") folderId: Long): NetworkResponse<FolderInfoResponse>
 
     // 폴더 내 게시글
     @GET("/api/v1/folders/{folderId}")
-    suspend fun requestDetailListFolder(@Path("folderId") folderId: Int, @Query("end") end: Int): NetworkResponse<DetailFolderResponse>
+    suspend fun requestDetailListFolder(@Path("folderId") folderId: Long, @Query("end") end: Int): NetworkResponse<DetailFolderResponse>
 }
